requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
facebook-business==18.0.0
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
cryptography==41.0.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0 